Refactor JSONMessageRegistryTransport into multiple files
